It works!!! It fucking works!
The last build spat-out my first CI-built AppImage of a kivy app! Note
that it's 47M just for a dumb Hello World GUI app. So, yeah, trade-offs..

I added a flowerbox to the build script with some links to the docs that were critical in helping me make this possible. I'll re-add them to this commit too:

 * https://github.com/AppImage/AppImageKit/wiki/Bundling-Python-apps
 * https://kivy.org/doc/stable/installation/installation-linux.html
 * https://kivy.org/doc/stable/guide/basic.html

And here's a link to the first successful CI job. Yeah! :D

 * https://gitlab.com/maltfield/cross-platform-python-gui/-/jobs/573939488