damn, so AppImage execution is failing to run in GitHub's shared runner with fuse errors. This is most likely because fuse requires a kernel module to be loaded, and for that to work in a Docker container, then the docker host must have that module loaded.
  2020-05-31T21:33:35.5478138Z + chmod +x /tmp/appimagetool.AppImage
  2020-05-31T21:33:35.5487678Z + mkdir dist
  2020-05-31T21:33:35.5517125Z + /tmp/appimagetool.AppImage /tmp/kivy_appdir dist/helloWorld.AppImage
  2020-05-31T21:33:35.5522850Z fuse: device not found, try 'modprobe fuse' first
  2020-05-31T21:33:35.7058092Z open dir error: No such file or directory
  2020-05-31T21:33:35.7058725Z + uname -a

The question is: am I running a container in a container? Because if the ubuntu docker host is actually a VM with debain running inside it as a container, then maybe I can load the fuse module on the docker host before going into the container? Let's try it.

fwiw, this is not necessary on gitlab's shared runners. I also found this indicating that github had no desire to add modules totheir shared runners :\

 * https://github.community/t/custom-linux-kernel-module/17194

And I learned that apparently the GitHub MacOS shared runners support nested virtualization and have vagrant & virtualbox installed by default. Intersting..

 * https://github.com/actions/virtual-environments/issues/183
 * https://github.com/actions/virtual-environments/issues/433#issuecomment-610789144