cool, the last run confirmed that our ubuntu host can run AppImages and fuse is not an issue
 * https://github.com/maltfield/cross-platform-python-gui/runs/725641922?check_suite_focus=true

Let's see if I can figure out how to tell it to use the docker image for a single step only *after* we run a command on the docker host (ubuntu) to setup fuse. Oh, we'll also want to checkout the code before the script runs -- both actions inside the docker container. Possible?

If this is too hard, I'll try harcoding the git clone in my script.

Worst case, I'll abandon docker and just do everything directly in ubuntu.