the last run didn't produce much info because something in the process of uplaoding and running the mac/buildDmg.sh script is replacing the final two characters = '\;' in my `find` command with ';' (removing the backslash). Anyway, let's just try xargs.
Also, I noticed that `python3` (which we don't call, but may be called by one of the kivy scripts) is actually /usr/local/bin/python3 which is a symlink to /usr/local/bin/Cellar/python/3.7.7/bin/python3.7 which is installed by brew and is version 3.7.7. But /usr/bin/python3 is the system's default python, which is 3.7.3. So I'm also declaring an alias in hope that subsequant comamnds will use it. If that fails, maybe I'll try to create a virtualenv