added the asynchronous structure for setting up a usb "hotplug callback" that calls our defined function when a usb event occurs (ie: removing a usb drive) using the python libusb implentation in module = libusb1.
There's several libusb solutions for python, but libusb1 appears to be the smallest and most mature. And it does support registering a callback function that's called when an event (like a usb device being removed) is triggered

 * https://github.com/vpelletier/python-libusb1/blob/master/examples/hotplug.py
 * https://kivy.org/doc/stable/guide/widgets.html

Unfortunately, registering the callback is blocking, so that froze the UI when calling toggle() at first (when the user clicks the "arm (buskill)" button)

I tried using asyncio and threadding before deciding on multiprocessing. I chose multiprocessing mainly because it allows me to terminate the child process, something not easily done with threads using the threadding module. I probably shouldn't be just terminating the process, but libusb1's USBContext.handleEvents() function (which is the call that waits for the usb event and calls our function) doesn't appear to have a way to tell it to exit safely (other than USBContext.close()--which can hang for up to <timeout>).

An alternative to terminate() might be to set the timeout of handleEvents (which defaults to 60 seconds) down to 1 second or something, but the UI would still lock-up for at-most 1 second with that solution. I could also do 1/60th of a second, but I don't know what the consequences of calling handleEvents() 60-times-per-second would be. Possibly nothing, but terminate() just seems like a better solution.

I haven't actually tested any hotplugging events yet; that comes next.

I have no idea yet if this will actually work with permissions (ie: not require root) or if it will work on windows/macos (I've read/heard there's some issues with libusb on those platforms). If libusb requries root to register a callback on usb hotplugging events, then I might just abandon this and go back to manipulating udev files, which is probably faster/safer than this libusb approach. We'll see..

Oh, I also made a bunch of subtle changes to the GUI layout, which make it much more sane when the window is resized