I finished implementing the lockscreen trigger in linux with this commit!
I tested that this works with Ubuntu 20.04. Happily, I was able to get immediate hotplug notifications on usb device insert and removal (though we only care about removal) as an unprivliged user, which called my trigger() function with some basic info about the usb device and the event (and USBContext for further device investigation).

This is also fully functional in the CLI version by calling `buskill --arm` or `buskill -a`

I found that hitting the escape button closed the window, which could too-easily disarm buskill undesireably--a dangerous default behaviour. I fixed that in this commit.

I also found a bug where the GUI wouldn't close if BusKill was aremd--because it was waiting for the child process. I fixed this by setting up the kivy window to call a close() function which reaches back to buskill.close() which terminates the buskill_handler child process. Whew! It works :)

Unrelated: when I switched to Ubuntu 20.04 for this test (which has python2.8), I realized that my AppImage for linux is totally broken; the python binaries were just symlinks back to /usr/bin/pyton3.7. Fix coming soon.