this is an untested first commit attempting to detect usb removal events in windows using the pypiwin32 modules
 * https://pypi.org/project/WMI/

Unfortunately, it appears that libusb (which is what we use to implement the callback registration of usb hotplugging events for this app in linux) does not implement hotplugging events notifcation in Windows, and an issue has been open for >5 years requesting this feature.

 * https://github.com/libusb/libusb/issues/86

I was successfully able to get usb removal detection working by using this script:

 * https://stackoverflow.com/questions/38689090/detect-media-insertion-on-windows-in-python

This commit tries to integrate the code from the above SO question into our code-base; it's untested.