attempting to try to use my linux libusb implementation on macos. this commit fi xes the following:
1. the multiprocessing module spammed me with this error, which I fixed by expli
citly setting it to spawn in main.py

  The process has forked and you cannot use this CoreFoundation functionality safely. You MUST exec().
  Break on __THE_PROCESS_HAS_FORKED_AND_YOU_CANNOT_USE_THIS_COREFOUNDATION_FUNCTIONALITY___YOU_MUST_EXEC__() to debug.

2. libusb needs to be installed on MacOS. I had to do this on Windows too--befor
e I abandoned libusb as an option because it lacks hotplugging support in Winodw
s. It seems sufficient to copy this libusb-1.0.dylib in the same dir as the othe
r python source files. We'll see if that plays nice with PyInstaller

So far I was able to successfully get my linux code to register my callback function on hotplugging events on my headless Mac Mini, and it gave me a list of devices as "inserts" when that's first setup, as is the normal functionality with Linux too. Unfortunately I can't test this since I don't physically have a mac, and I don't think I can pester the fine folks at MacStadium to insert and remove a USB drive over-and-over for me to test this :)

If this builds with the libusb depends successfully, then I can pass it to an actual Macbook owner to test.