the last build was functional on all 3 platforms!
This is the first commit where I can say that this app is finally functional on all 3 platforms: Linux, Windows, and MacOS. Further testing will be done, but I won't be implmeneting new features until our first release of v0.1.0 is out

Before I can make a release, we need to address some prerequisite tasks around documentation and security:

 1. I need a proper documentation solution. I'll probably use Sphinx / Read The Docs for this

 2. I need to sign all my commits on github with gpg

 3. I need to update the build scripts to verify the integrity of any depends being downloaded

 4. I need to make sure that the builds are deterministic. Any files produced by the build from the GitHub's CI process should match exactly the files produced by the bulild from my local machine.

 5. I need to find out if it's possible for me to upload a (distinct, purpose-created) pgp private key to GitHub somehow such that it's secure/not-public and can be used by the CI process to cryptographically sign build artifacts

 6. I need to setup TUF to sign releases. This will also be leveraged in a later release when I implement a self-updater feature into this app.

I hope that this commit solves #2 above, as I've added a signing subkey to my primary, master gpg key and copied it to my development environment, uploaded my updated public key to my GitHub account, and enabled commit signing by default in this repo's remote sandbox on my machine. Hopefully this commit will be signed and GitHub will give it a green checkbox.

As part of addressing #3, my build process currently requires downloading two dependencies from pip. Pip does not provide cryptographic authentication or integrity checks, but it does have a system for project maintainers to sign their releases with gpg and upload that to PyPI using twine. I've opened tickets requesting that the developers for the projects that we depend on begin to sign their releases to PyPI so that our build process can be more secure. In the meantime, I should 3TOFU a set of checksums and update the build scripts to fail if the integrity doesn't match the hard-coded checkums.

 * https://github.com/vpelletier/python-libusb1/issues/54
 * https://github.com/kivy/kivy/issues/6973