updated appimagetool to be deterministic for reproducable builds
 * https://github.com/BusKill/buskill-app/issues/3

Unfortunately the latest stable version of appimagetool (v12 from 2019-05) uses a very old version of mksquashfs (v4.3 from 2014-05) which does not support building reproducable builds due to some parallelization optimizations

 * https://github.com/AppImage/AppImageKit/releases
 * https://sourceforge.net/p/squashfs/code/ci/master/tree/CHANGES
 * https://github.com/AppImage/AppImageKit/issues/929#issuecomment-653641526

Other projects like the electurm crypto wallet got around this issue by hacking the latest stable appimagetool to use a fork of squashfs-tools = squashfskit

 * https://github.com/squashfskit/squashfskit
 * https://github.com/SomberNight/electrum/blob/ae714772c38410a0169f2c76a14a64a62c0daff0/contrib/build-linux/appimage/build.sh#L212-L225

But in 2019-08, squashfs-tools released an updated version (v4.4) which added support for reproducable builds, as well as fixed a couple CVEs and other changes

 * https://lore.kernel.org/lkml/CAB3wodcL=gnQOmHGGNukWK3OUbU2p=OHzLmzPi7ns_WNTGBEwg@mail.gmail.com/

Therefore, in this commit, I've added the code to make changes to appimagetool's latest release to use the updated version of squashfs-tools.

The asked the maintainer of AppImageKit about this, and they said there's no future releases expected for the existing appimagetools repo, as development has been moved to a Go version, which is currently still a work-in-progress with no eta on it becoming stableo

 * https://github.com/SomberNight/electrum/blob/ae714772c38410a0169f2c76a14a64a62c0daff0/contrib/build-linux/appimage/build.sh#L212-L225

So, for now, this hack is as good as it gets.