fixed path to /bin/sh. The last build failed with
  2020-07-04T17:36:43.2225576Z build/linux/buildAppImage.sh: /tmp/appimagetool_appdir/usr/lib/appimagekit/mksquashfs: /usr/bin/sh: bad interpreter: No such file or directory

In debian, /bin is just a symlink to /usr/bin, which is where things like 'sh' and 'bash' live. I guess in Ubuntu that's not the case. Surprising, since Ubuntu is based on Debian.

I wish I could just run this in docker, but I think I had some issue with github workflows when I initially set this up that made docker not an option..