forcing install of pip depends, even if already installed outside the AppDir
The last build's checksum was different than my local build. I extracted both AppImages and found the diffs were related to the libusb1 package being installed in the AppImage from GitHub, but not on my local build

  user@buskill:~/sandbox/buskill-app/dist$ diff -r squashfs-root/ ~/Downloads/dist/squashfs-root
  Only in /home/user/Downloads/dist/squashfs-root/opt/python3.7/lib/python3.7/site-packages: libusb1-1.8.dist-info
  Only in /home/user/Downloads/dist/squashfs-root/opt/python3.7/lib/python3.7/site-packages: libusb1.py
  Only in /home/user/Downloads/dist/squashfs-root/opt/python3.7/lib/python3.7/site-packages: usb1
  Only in squashfs-root/opt/src: buskill.log

Further investigation into my local build's log showed that it skipped the libusb1 install because it found it on my dev machine in my user's home dir, outside the AppDir. I'm fixing this in this commit so it force-installs it into the AppDir by adding the --ignore-installed arg

 * https://stackoverflow.com/questions/2915471/install-a-python-package-into-a-different-directory-using-pip