adding all the dependencies and their checksums. And updated linux build script to use them. For more info, see:
 * https://github.com/BusKill/buskill-app/issues/2

My initial plan was to and add a checksum validation after the downloads, but I ended up deciding to keep all the dependency assets in the repo when I hit a 404 on one of the dependencies. There's pros and cons to both, but I think it's more stable long-term to have all the assets already in the repo and not depend on any external downloads.

I'm also adding firejail to wrap the pip install commands just to make damn sure it doesn't fetch anything from the internet. Pip doesn't do a cryptographic integrity check against its downloads, which is the main reason I'm making these changes to being with--to make sure our builds aren't poisoned by a MITM attack at build-time.