beginning to implement pre-release signing
 * https://github.com/BusKill/buskill-app/issues/4

In this commit I'm attempting to add a new workflow that's triggered every time a release is published. I'd like to archtect it this way so that I don't have to implement gpg key signing in all three platforms; better that it's a distinct job that can just download the assets produced by the build job (even from windows/MacOS builds) and sign them using linux, which is more objectively more secure (apt) and subjectively easier to use than MacOS & Windows.

Unfortuantely it appears that we can't just trigger one job to execute another job. It's possible, but first I'm going to see if I can hook into the release. But writing a workflow triggered by a release that edits a release risks creating an infinite loop, so here I'm just going to try to get the scope of the hook correct.

 * https://github.community/t/inability-to-trigger-workflows-from-within-a-workflow/17597
 * https://github.com/mvasigh/dispatch-action