changing the "sign pre-release" workflow to be triggered by a custom dispatch
Yeah, apparently it's not possible to have a workflow trigger another workflow, which sucks

 * https://github.community/t/github-actions-workflow-not-triggering-with-tag-push/17053

In fact, the creds used in GitHub Actions don't have the permission to launch any workflow, so that's the bottom-line issue. The solution for this fail is to create a PAT = Personal Access Token, make it available to the GitHub action as a secret (environment varialbe) and then use *that* cred to trigger a workflow manually via the repository_dispatch API call.

To this end, I created a "buskill-bot" GitHub user:

 * https://github.com/buskill-bot

It's not very well documented, but I found that the minimum necessary permissions for a user to be able to initiate "repository_dispatch' calls to the API requires that the user to have write permission to the given repo and the PAT needs exactly 1 scope checkbox ticked when creating it = 'public_repo' (under the "repo" section at the top). Note that I had to explicitly go to the "'buskill-app' repo -> settings -> Manage Access" and change the "BusKill Bot" permissions to "Write" as they default to "Read". Triage was too low, Maintain was not necessary (too high).

If you don't do this, the GitHub API doesn't give you a helpful error; they just give you a 404 (not even a 403 or *something* helpful!). Anyway, I finally got a 204 which is actually a success message when I got these min permissions dialed-in:

  user@disp1246:~$ curl -iu buskill-bot:${token} -d '{"event_type":"event_type"}' https://api.github.com/repos/buskill/buskill-app/dispatches
  HTTP/1.1 204 No Content
  Server: GitHub.com
  Date: Sat, 11 Jul 2020 17:17:05 GMT
  Status: 204 No Content
  X-RateLimit-Limit: 5000
  X-RateLimit-Remaining: 4972
  X-RateLimit-Reset: 1594489638
  X-OAuth-Scopes: public_repo
  X-Accepted-OAuth-Scopes:
  X-GitHub-Media-Type: github.v3; param=everest-preview; format=json
  Access-Control-Expose-Headers: ETag, Link, Location, Retry-After, X-GitHub-OTP, X-RateLimit-Limit, X-RateLimit-Remaining, X-RateLimit-Reset, X-OAuth-Scopes, X-Accepted-OAuth-Scopes, X-Poll-Interval, X-GitHub-Media-Type, Deprecation, Sunset
  Access-Control-Allow-Origin: *
  Strict-Transport-Security: max-age=31536000; includeSubdomains; preload
  X-Frame-Options: deny
  X-Content-Type-Options: nosniff
  X-XSS-Protection: 1; mode=block
  Referrer-Policy: origin-when-cross-origin, strict-origin-when-cross-origin
  Content-Security-Policy: default-src 'none'
  Vary: Accept-Encoding, Accept, X-Requested-With
    X-GitHub-Request-Id: 1AAB:5933:1439A5C:222A5C6:5F09F411

  user@disp1246:~$

This commit will hopefully register the "sign_workflow" workflow to be executed when any repository_dispatch is called, like the example command above