adding curl
I was successfully able to limit the execution of the sign_workflow to when the 'event_type' is equal to 'sign_release' in the last commit. I was also alble to successfully send the client_payland, validate it, and set it to an environment variable in the workflow.

First I got this when I exectued the dispatch without any payload (so the payload and the release_id were both null)

  user@disp1246:~$ curl -iu buskill-bot:$token -d '{"event_type":"sign_release"}' https://api.github.com/repos/buskill/buskill-app/dispatches

 * https://github.com/BusKill/buskill-app/runs/861373172

Then I got this when I executed the dispatch with a release_id of '123'. I got an error that our debian-slim docker image didn't have curl apparently, so this commit iterates by adding curl with apt..

  user@disp1246:~$ curl -iu buskill-bot:$token -d '{"event_type":"sign_release", "client_payload": {"release_id":"123"}}' https://api.github.com/repos/buskill/buskill-app/dispatches

 * https://github.com/BusKill/buskill-app/runs/861376212