attempting better input sanitization
Unfortunately, even my last commit is vulnerable to arbitrary command injection if the user adds a single quote (') instead of a double quote (").

  user@disp1246:~$ curl -iu buskill-bot:$token -d '{"event_type":"sign_release", "client_payload": {"release_id":"28469791'"'"'`echo fail`"}}' https://api.github.com/repos/buskill/buskill-app/dispatches

And the site I referenced in a previous commit for how they sanitize inputs? Well, I proved that they're actually vunlerable to an arbitrary code injection attack too :/

 * https://github.com/victoriadrake/github-guestbook/issues/1#issuecomment-657121754
 * https://victoria.dev/github-guestbook/

Because github's workflows esentially write the string as a literal directly into the script (I guess?) rather than storing items in the 'github' context as pre-existing variables, it seems that we can't solve this with bash. Fortunately, I discovered that there's GitHub has some functions we can use. So before we make it accessible to bash, we can first exit if it contains a single apostrophie. If it doesn't then it should be safe to assign it to a variable using bash variable expansion to meet our strict sanitization allowlist/regex without single quotes in the github value breaking this expanison procedure