adding 'docs' dir for sphinx-built documentation
  * https://www.sphinx-doc.org/en/master/usage/configuration.html
  * https://www.writethedocs.org/guide/tools/sphinx/

Eventually this will end up on docs.buskill.in, likely as a GitHub page. Or ReadTheDocs.

Here's a couple brilliant sphinx-powered sites that I took inspiration from.

 * https://docs.phpmyadmin.net/en/latest/
 * https://zulip.readthedocs.io/en/latest/
 * https://ryan-roemer.github.io/sphinx-bootstrap-theme/#
 * https://cassandra.apache.org/doc/latest/
 * https://book.cakephp.org/4/en/index.html

More examples:

 * https://www.sphinx-doc.org/en/master/examples.html

We're going to use reStrutredText, as it seems like the best choice for being able to do advanced things like embedding a Bill Of Materials spreadsheet and embedding videos that fall-back to a URL/QR Code when rendered to PDF for printing to paper (or epub?)

 * https://softwarerecs.stackexchange.com/questions/75204/documentation-as-code-markup-supporting-tables-with-calculations-basic-spreads
 * https://stackoverflow.com/questions/62682095/how-to-add-a-spreadsheet-in-read-the-docs/62683492#62683492
 * https://stackoverflow.com/questions/62682412/video-fall-back-when-printing-to-paper-with-read-the-docs
 * https://github.com/brechtm/rinohtype/issues/172