adding non-functonal script to build documentation using sphinx
I was expecting that the GitHub API exposed some option to upload some files or artifact that would upate the GitHub Page(s), but that's not how it works. Apparently I the contents that becomes the GitHub Page(s) must either be in the root of the 'master' branch, a '/docs' dir in the master branch, or the root of a branch named 'gh-pages'.

All of the examples I found that built a static site using GitHub Actions and pushed those changes to a GithHub Page  actually just use a secret token to git clone, build, then overwrite the contents of the 'gh-pages' branch, and commit/push those changes. I guess that's what we'll do

 * https://developer.github.com/v3/repos/pages/#update-information-about-a-github-pages-site
 * https://github.com/ammaraskar/sphinx-action
 * https://github.com/shalzz/zola-deploy-action/blob/master/entrypoint.sh
 * https://github.com/khanhicetea/gh-actions-hugo-deploy-gh-pages/blob/master/entrypoint.sh
 * https://github.com/BryanSchuetz/jekyll-deploy-gh-pages/blob/master/entrypoint.sh
 * https://github.com/bocoup/wpt/blob/9b5e2cd9e38be5b6c765e157fd9e00c42efd31f6/tools/ci/website_build.sh

That doesn't answer the question about how we get versioned documentation that cooresponds to our tagged versions in github. The answer appears to be this sphinxcontrib-versioning extension for sphinx. It has a lot of great features that should allow me to tune it to do exactly what I want. But I'll deal with that later..

 * https://sphinxcontrib-versioning.readthedocs.io/en/latest/tutorial.html

This commit just sets-up the workflow and its script and prints some env vars that I'll hopefully be able to use to do the git checkouts more robustly