finished most of the initial structure of the docs
I also setup updated our sourcecode with the most basic docstrings. Eventually I'll need to document the functions/classes, but now it's just per-file. This change plays very well with and looks great in Sphinx.

I tried to setup i18n/languages, but the URL still doesn't end with '/en/latest' as most Sphinx sites do, and I couldn't find any documentation on how to set that up, so I created a SO question and moved-on. This is important to setup properly before we launch (even before we actually translate to another language) so that the URLs permalinks won't change.

 * https://stackoverflow.com/questions/62884907/how-to-internationalize-sphinx-i18n

I'm a terrible speller and usually use vim for all my editing. So, protip note to self: executing the following command from the `docs/` dir opens all the .rst files in gvim with one-file-per-tab.

  find . -type f -name "*.rst" | xargs gvim -p

This can be done with vim, but I'd rather right-click than type 'z=' on a word for spell-check. If you use the following in your .vimrc, then you can switch between tabs with F1 (go back one tab = to the left) and F2 (go forward one tab = to the right). And you can right-click on a misspelled word (after left-clicking on it) to choose from suggestions

  set ts=3
  syntax on
  colorscheme torte
  set backspace=indent,eol,start
  :nmap <F1> :tabp<CR>
  :nmap <F2> :tabn<CR>
  :let @c='^[yyjpI^[lr*A^['
  :set ruler
  :set spell
  :set mousemodel=popup
  :set tabpagemax=99

Lots more quality documentation coming soon (hopefully without spelling mistakes)