added images and emojis
Emojis can be added to sphinx by directly adding the unicode to the .rst files, but that maakes editing difficult on editors that don't support emojis, so it's not ideal. It's better if we can type :smile: like we can with github markdown.

Soolution: I'm adding this Sphinx extension = sphinxemojicodes. You can get a list of the supported emojis and their codes (ie: |:smile:|) here:

 * https://sphinxemojicodes.readthedocs.io/en/stable/#supported-codes