changing symlink to text file with markdown link
Symlink work fine in git, but the UX on GitHub's WUI is less-than-ideal: clicking the 'KEYS' file in the last commit just opened the "file" as if it were a text document with the contents = "docs/security/pgpkeys.rst".

I was hoping it would just follow the symlink and open the file that it points to in the repo, but it doesn't. It didn't even provide any way for the user to click and resolve the symlink!

So I'm going to try to provide that "click" mechanism manually by changing from a symlink to a dumb markdown with a link in it with this commit.