modified the rtd theme template and the sphinx config and the updatePages.sh script to add the lower-left rtd menu for users to change versions and languages and download alternate formats like PDFs and epbubs. This is not functional yet.
I found the whole process to host sphinx/Read the Docs on GitHub pages to be so annoyingly opaque that I published this guide on how to do it yesterday:

 * https://tech.michaelaltfield.net/2020/07/18/sphinx-rtd-github-pages-1/

I have to do it in two-parts because just getting the most basic GitHub Pages-powered sphinx site (with the Read the Docs theme) built by GitHub Actions setup took a whole lot of explaining.

Now I'm going to use this repo as an experiment to hack out a solution for internationalization/versions/downloads and then I'll merge it into my generic fork-ready helper-repo here:

 * https://github.com/maltfield/rtd-github-pages/tree/master

Anyway, this commit achieves a lot of the work on getting that lower-left rtd menu working, but I'm committing now before I finish it because I have to create a second branch so I can test the "versions" bit in the lower-left menu.

I'd like to follow a "git flow" style branching system where "master" is the stable code that's never committed-to directly (only merged-to) and a 'dev' branch where we commit everything-to. Not to mention feature and release and hotfix branches.

Anyway, I wanted to commit my changes just before I create the 'dev' branch. Once I do, I'll *only* be committing to 'dev'.