adding bill of materials
In this commit, I'm including my first sphinx extension, which I found was necessary in order to use python logic to conditionally have a hyperlink point to one of two locations depending on the buildername.

 * https://stackoverflow.com/questions/63099885/variable-external-links-based-on-sphinx-build-destination-eg-html-vs-pdf/63132102#63132102

I want to be very clear: BusKill uses affiliate links to fund its development as an open-source project. Currently the only affiliate program we're a member of is Amazon, but I'm open to extending that in the future. Amazon was just the easiest. As such, if someone clicks links on our websites to an amazon product (ie: a usb drive or extension cable) and buys it, we make a small commission from that sale.

Amazon has a ton of restrictions on their affiliate program. One is that you can't add their links to any offline medium or via emails. Therefore, since our sphinx documentation was explicitly designed to be able to render to both online html (in which we're allowed to put affiliate links) and offline PDFs (in which we're not allowed to add affiliate links), this commit to choose which link to add based on the buildername was created.

Therefore, the canonical URL where users should be linked-to to access our affiliate links (via email or whatever) is the following Bill of Materials page:

 * https://docs.buskill.in/buskill-app/en/stable/hardware_dev/bom.html