added gnupg support to verify the signature of the checksum file
I originally wanted to have the gpghomedir in ram-only, but since the python-gnupg module essentially just shells-out to the `gpg` command, it wouldn't be able to access the python in-memory filesystem. Instead, I just setup a '.buskill' dir that's only accessible to the local user running the BusKill app and add an ephemeral 'cache' dir inside of that where we download our files and create our .gnupg home dir.

Now the question is: how portable is this?