adding fall-back for the DATA_DIR under which our app has write access to store some cache files (and eventually our config file).
The last build failed for linux because I was trying to create a dir named '.buskill' in the same dir where 'main.py' lives, but this doesn't work with the AppImage because that location is a read-only squashfs.

This commit adds some fallback functionality. It will choose the first dir in a list of dirs that it can confirm write access to. First it tries the dir where 'main.py' lives. Next it tries the dir where AppImage lives (actually where sys.executable lives, which might be 'buskill.exe' or 'python3' -- we don't use it if it's 'python3' so we avoid using '/usr/bin/' or something). Finally, it tries '$HOME/.buskill/' in what is probably some OS-agnostic way