attemtping to fix upload
The last build failed because the ${ARCHIVE_DIR} variable wasn't known to the "upload-release-asset@v1" action.

	2020-08-21T11:33:31.6764792Z ##[group]Run actions/upload-release-asset@v1
	2020-08-21T11:33:31.6764928Z with:
	2020-08-21T11:33:31.6765086Z   upload_url: https://uploads.github.com/repos/BusKill/buskill-app/releases/29976726/assets{?name,label}
	2020-08-21T11:33:31.6765416Z   asset_path: dist/buskill-lin-x86_64.218296398.tbz
	2020-08-21T11:33:31.6765524Z   asset_name: ${ARCHIVE_DIR}.tbz
	2020-08-21T11:33:31.6765628Z   asset_content_type: application/x-bzip2
	2020-08-21T11:33:31.6765948Z env:
	2020-08-21T11:33:31.6766653Z   GITHUB_TOKEN: ***
	2020-08-21T11:33:31.6766765Z ##[endgroup]
	2020-08-21T11:33:32.4553010Z ##[error]ENOENT: no such file or directory, stat 'dist/buskill-lin-x86_64.218296398.tbz'

I don't have a good way to maintain variables between the steps, especially with these damn generic actions that black-box my steps. Black boxing is fine for really complex stuff, but this is only a few lines--why bother?

This commit attempts to take what I learned from the "sign" workflow and apply it to the "build" workflow -- we'll just do the upload ourselves so that we can use BASH to get the name of the tarball