added the logic for the upgrade() install in Windows & MacOS
Neither of these OS upgrades have been tested yet--so far only linux.

Note that I'm not using a python module with MacOS because to do so would require to bundle our app with a ton of pip modules, which decreases the security of the app. Instead, I just subprocess call out to `hdiutil` -- which itself has no way to just "extract" the contents of the image (actually a 7zip archive), so I mount it, copy it, and unmount.

In theory, anyway. Testing & fixes to come..