got the hamburger menu to actually open the nav bar
so this is mostly working. At least, it's opening the nav drawer when the hamburger menu is clicked.

But, since I'm actually not going to have any icons in the ActionBar, I'm going to go back to MDToolbar since it looks better, even if its ActionOverflow is totally missing

 * https://stackoverflow.com/questions/63655339/actionoverflow-actionbar-equivalent-with-kivymd-toolbar
 * https://github.com/BusKill/buskill-app/issues/6#issuecomment-683708088

Hopefully they'll fix this upstream before I need to add any items to the top MDToolbar. Else I'll switch back to ActionBar or have to write it myself

 * https://github.com/kivymd/KivyMD/issues/499