I'm abandoning KivyMD.
It was indeed pretty beta. Lots of the tutorials I followed online were out-of-date due the API changing. And a lot of functionality defined in the Google MD spec were missing. But that was OK.

I also encountered a lot of bugs, but--still--I value the UX, so I was still planning to use KivyMD despite all this. I figured that I would just report the bugs I found, maybe contribute a few PRs to fix the issue, and--over time--KivyMD would mature along with our app. Hey, it does appear to be actively developed. That's good.

My first few bug reports to [KivyMD](https://github.com/kivymd/KivyMD/issues/) over the past couple days went well. The maintainers responded positively within 24 hours. Then today I spent a couple hours reporting two more bugs. It took at least an hour to distill-down one of them into a short-as-possible code snippet (with only 2 lines modified from their example code copied straight from their reference documentation--isolating the cause of the bug to hopefully make their tracing easier).

I then typed-up a nice set of "steps to reproduce" commands that clearly demonstrates the bug. Anyone running debian (or in a VM) could reproduce it very easily--just by copying & pasting.

 * https://github.com/kivymd/KivyMD/issues/502

Within 30 minutes, the bug was picked-up. That's good.

But it was immediately closed by the maintainer, who said (in many words with a lot of shouting/exclamation marks) that it works fine. But they weren't testing against the code I provided (again, which was just 2 lines changed from their doc's example code). And they weren't even executing it on the same OS (which I made sure to provide). They weren't even running the same version of KivyMD. And they certainly didn't follow my "steps to reproduce" commands *at all*.

I asked politely to re-open the ticket and to follow my steps to reproduce. I got a lot more shouting back at me. Eventually another dev came-in and confirmed the bug and that it has been recently fixed in `master`.

It's one thing for an open-source project to be in beta. It's another thing to have toxic maintainers that won't work with contributors to fix their bugs that are carefully reported by the community.

This is not something I want to integrate into our BusKill app.