completely removed KivyMD
I've mostly replicated the "feel" of what I had with KivyMD in native kivy. Note that I'm importing the Roboto google fonts that came with KivyMD into this repo now. Roboto and MaterialIcons are the only fonts that are used by KivyMD.

I didn't copy the MaterialIcons fonts from KivyMD (which was a file named 'fonts/materialdesignicons-webfont.ttf'). Instead, I downloaded the 'MaterialIcons-Regular.ttf' file from here

 * https://github.com/google/material-design-icons/tree/master/font

One benefit is that I was also able to download 'MaterialIcons-Regular.codepoints' as well from the above link, which tells the unicode codepoints for each character. The characters names can be matched from here:

 * https://material.io/resources/icons/?icon=menu&style=baseline

This can be done with "[font=mdicons]\ue5d2[/font]" or similar.

I did try to get the hamburger menu as a character rather than an image, but the way ActionBar works in kivy, I wouldn't be able to put a character to the left of the text on the bar. The only way to do this was with the image icon for ActionPrevious, so I downloaded the png of the 'menu' icon from the above link and just made a few basic changes until it looked like the MDToolbar one.

There are a few remaining side-by-side differences between the previous commit and this one:

 1. KivyMD has a slightly taller MDToolbar than our ActionBar. The fonts and menubar are the same height, but KivyMD has more top & bottom margin

 2. The background color in KivyMD's label is grey. This commit's is black.

 3. There's no shadow under the ActionBar. This could be because the bg is black or because it's not something native Kivy does.

 4. I no longer have a navigation drawer

 5. I no longer have an update button/icon in the nav drawer

 6. I no longer have a dialog that opens when the above button in the nav drawer is clicked

I'll build the old functionality back sans-kivymd in the next commits