finished all 3 modals of the upgrade() process via GUI
The user can now do the following to download the latest version of the app in the GUI:

 1. Click the hamburger menu in the top-left to open the navigation drawer

 2. Click the "Update" button in the navigation drawer

 3. Click "Check Updates" in the confirmation dialog that appears

 4. Wait for some time on the interstitial dialog window (displaying an android-like spinning, indeterminate progress circle) while the upgrade() function runs, downloads files and verifies their signatures, and extracts them into the EXE_DIR

 5. After the upgrade succeeds (or fails), a final dialog replaces the interstitial dialog with the result of the upgrade()

I don't yet have the "Restart Now" button on the final confirmation dialog doing anything.

I should also test that pressing the "Cancel" button on the interstitial dialog while upgrade() is running will actually kill the process and clean the Pool(). Ah, and I should make sure to close() the pool when its process is ready()!