failed to fix the 'cancel' button during upgrade()
I added the code needed to terminate the pool (and its processes) when the "cancel" button is clicked on the interstitial dialog while upgrade() is running. Unfortunately, terminate() appears to be blocked and doesn't finish until after upgrade() finishes running.

This isn't what should happen. I can't understand it, but it does happen. Instead I'm going to switch back to multiprocessing.Pipe() directly and using a multiprocessing.Pipe() to get the return value (and maybe even the upgrade_status_msg too)